add: Added verify events
