+++ src/events/CommandEvents/autoRoleEvent.js
